move setNodeValue and setNodePositionById
