+++ packages/frontend/editor-ui/src/stores/workflows.store.test.ts
@@ -1030,51 +1030,6 @@ describe('useWorkflowsStore', () => {
 		});
 	});
 
-	describe('setNodeValue()', () => {
-		it('should update a node', () => {
-			const nodeName = 'Edit Fields';
-			workflowsStore.addNode({
-				parameters: {},
-				id: '554c7ff4-7ee2-407c-8931-e34234c5056a',
-				name: nodeName,
-				type: 'n8n-nodes-base.set',
-				position: [680, 180],
-				typeVersion: 3.4,
-			});
-
-			expect(workflowsStore.nodeMetadata[nodeName].parametersLastUpdatedAt).toBe(undefined);
-
-			workflowsStore.setNodeValue({ name: 'Edit Fields', key: 'executeOnce', value: true });
-
-			expect(workflowsStore.workflow.nodes[0].executeOnce).toBe(true);
-			expect(workflowsStore.nodeMetadata[nodeName].parametersLastUpdatedAt).toEqual(
-				expect.any(Number),
-			);
-		});
-	});
-
-	describe('setNodePositionById()', () => {
-		it('should NOT update parametersLastUpdatedAt', () => {
-			const nodeName = 'Edit Fields';
-			const nodeId = '554c7ff4-7ee2-407c-8931-e34234c5056a';
-			workflowsStore.addNode({
-				parameters: {},
-				id: nodeId,
-				name: nodeName,
-				type: 'n8n-nodes-base.set',
-				position: [680, 180],
-				typeVersion: 3.4,
-			});
-
-			expect(workflowsStore.nodeMetadata[nodeName].parametersLastUpdatedAt).toBe(undefined);
-
-			workflowsStore.setNodePositionById(nodeId, [0, 0]);
-
-			expect(workflowsStore.workflow.nodes[0].position).toStrictEqual([0, 0]);
-			expect(workflowsStore.nodeMetadata[nodeName].parametersLastUpdatedAt).toBe(undefined);
-		});
-	});
-
 	describe('setNodes()', () => {
 		it('should transform credential-only nodes', () => {
 			const setNodeId = '1';
@@ -1596,7 +1551,7 @@ describe('useWorkflowsStore', () => {
 			await waitFor(() => expect(workflowsStore.selectedTriggerNodeName).toBe('n0'));
 			workflowsStore.removeNode(n0);
 			await waitFor(() => expect(workflowsStore.selectedTriggerNodeName).toBe('n1'));
-			workflowsStore.setNodeValue({ name: 'n1', key: 'disabled', value: true });
+			useWorkflowState().setNodeValue({ name: 'n1', key: 'disabled', value: true });
 			await waitFor(() => expect(workflowsStore.selectedTriggerNodeName).toBe(undefined));
 		});
 	});
